New car and pathway textures. Modification of gameobject structure for retieving static width and height related with a type - needed for editor.

+++ Auticka/src/cz/mff/cuni/autickax/entities/GameObject.java
@@ -30,7 +30,7 @@ abstract public class GameObject {
 
 	protected Autickax game;
 	protected GameScreen gameScreen;
-	protected byte type;
+	protected int type;
 
 	public GameObject(float startX, float startY, GameScreen gameScreen) {
 		this.position = new Vector2(startX, startY);
@@ -168,5 +168,5 @@ abstract public class GameObject {
 	 */
 	public float getRotation() {
 		return this.rotation;
-	}
+	}	
 }
